feat(tab5): implement index-based effectID indirection for hex-ID registry
The v2 firmware migrated from sequential integer effect indices (0-146) to
stable hex-based effectIDs (0x0100-0x1A09). The Tab5 encoder was completely
broken because all effectId fields were uint8_t, truncating values > 255.

Architecture: encoder scrolls through position indices (uint8_t 0-N) internally,
translated to/from hex effectIds (uint16_t) at the WebSocket boundary via
effectIdFromIndex() and indexFromEffectId() helpers.

- main.cpp: Add s_effectIds[] registry, translation helpers, position-indexed
  effects.list parser
- WebSocketClient.cpp: Translate index→hexId on send for effect and zone changes
- WsMessageRouter.h: Translate hexId→index on receive for zone status and sync
- ParameterHandler.cpp: Special effectId handling in applyStatus() for hex IDs
- ZoneComposerUI: Dynamic max from effect count, uint16_t ZoneState.effectId
- PresetData.h: V2 struct with uint16_t effectId fields (breaking change for
  saved presets, maintains 64-byte struct size)
- PresetManager.cpp: Translate between indices and hex IDs on save/load
- iOS: Comment-only updates (already compatible with hex IDs)

+++ firmware/Tab5.encoder/src/ui/ZoneComposerUI.h
@@ -21,7 +21,7 @@
  * Zone state for display
  */
 struct ZoneState {
-    uint8_t effectId = 0;
+    uint16_t effectId = 0;  // Actual hex effectId (e.g. 0x0100), NOT position index
     char effectName[48] = {0};
     uint8_t speed = 25;
     uint8_t paletteId = 0;
